refactor: Wave 6 quality push - god class decomposition, CI consolidation
God class decomposition (5 classes >1000 lines):
- FrankensteinEditor (1364 lines): extracted ClipboardMixin, HistoryMixin,
  TransformMixin + shared types module
- URDFTextEditor (1104 lines): extracted ValidationMixin, DiffMixin
  + shared types module
- GolfSimulationGUI (1255 lines): extracted StyleMixin, DockerMixin
- DoublePendulumApp (1053 lines): extracted PendulumRendererMixin
- MotionCapturePlotter (1418 lines): extracted mocap_data_loader module

CI workflow consolidation (49 -> 35 workflows):
- Removed 13 workflows using deprecated @google/jules npm CLI
- Removed 1 disabled workflow (Jules-Cleaner, if: false)
- Removed 1 redundant workflow (Jules-Consolidator, overlaps PR-Compiler)
- Updated Control-Tower and Manual-Run-All to remove dead references
