--- .github/workflows/Jules-Consolidator.yml
@@ -1,111 +1,0 @@
-name: Jules Consolidator (Daily PR Merge)
-
-on:
-  workflow_dispatch:
-    inputs:
-      dry_run:
-        description: "Dry run - analyze only"
-        required: false
-        default: "false"
-        type: boolean
-      exclude_prs:
-        description: "Comma-separated PR numbers to exclude"
-        required: false
-        default: ""
-        type: string
-  schedule:
-    - cron: "0 0 * * 0,4" # 5 AM PST Mon-Fri (13 UTC)
-
-concurrency:
-  group: jules-consolidator
-  cancel-in-progress: false
-
-permissions:
-  contents: write
-  pull-requests: write
-  issues: write
-
-jobs:
-  discover:
-    runs-on: ubuntu-latest
-    outputs:
-      pr_count: ${{ steps.list.outputs.count }}
-      pr_list: ${{ steps.list.outputs.prs }}
-      should_consolidate: ${{ steps.list.outputs.should_consolidate }}
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          fetch-depth: 0
-
-      - name: List Consolidatable PRs
-        id: list
-        env:
-          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          EXCLUDE_PRS: ${{ inputs.exclude_prs }}
-        run: |
-          prs=$(gh pr list --base main --state open --json number,title,author,labels,mergeable \
-            --jq '[.[] | select(
-              (.labels | map(.name) | contains(["wip"]) | not) and
-              (.labels | map(.name) | contains(["do-not-consolidate"]) | not) and
-              (.labels | map(.name) | contains(["blocked"]) | not) and
-              (.mergeable != "CONFLICTING")
-            )]')
-
-          if [ -n "$EXCLUDE_PRS" ]; then
-            for num in $(echo $EXCLUDE_PRS | tr ',' ' '); do
-              prs=$(echo "$prs" | jq --arg n "$num" '[.[] | select(.number != ($n | tonumber))]')
-            done
-          fi
-
-          count=$(echo "$prs" | jq 'length')
-          echo "count=$count" >> $GITHUB_OUTPUT
-          echo "prs=$(echo "$prs" | jq -c '.')" >> $GITHUB_OUTPUT
-
-          [ "$count" -gt 1 ] && echo "should_consolidate=true" >> $GITHUB_OUTPUT || echo "should_consolidate=false" >> $GITHUB_OUTPUT
-
-  consolidate:
-    needs: discover
-    if: needs.discover.outputs.should_consolidate == 'true' && inputs.dry_run != 'true'
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          fetch-depth: 0
-          token: ${{ secrets.GITHUB_TOKEN }}
-
-      - name: Configure Git
-        run: |
-          git config user.name "jules-bot"
-          git config user.email "jules-bot@users.noreply.github.com"
-
-      - name: Create and Merge PRs
-        env:
-          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          PR_LIST: ${{ needs.discover.outputs.pr_list }}
-        run: |
-          BRANCH_NAME="consolidated-prs-$(date +%Y-%m-%d)"
-          git checkout -b "$BRANCH_NAME" origin/main
-
-          included=""
-          for row in $(echo "$PR_LIST" | jq -r '.[] | @base64'); do
-            pr_num=$(echo ${row} | base64 --decode | jq -r '.number')
-            pr_title=$(echo ${row} | base64 --decode | jq -r '.title')
-
-            if gh pr checkout "$pr_num" --detach 2>/dev/null; then
-              git checkout "$BRANCH_NAME"
-              if git merge FETCH_HEAD --no-edit -m "Merge PR #$pr_num: $pr_title"; then
-                included="$included #$pr_num"
-              else
-                git merge --abort
-              fi
-            fi
-          done
-
-          git push origin "$BRANCH_NAME"
-
-          gh pr create \
-            --title "[Jules/Consolidator] Daily Consolidation - $(date +%Y-%m-%d)" \
-            --body "Consolidated PRs:$included" \
-            --label "jules:consolidation" \
-            --base main \
-            --head "$BRANCH_NAME"
