--- .github/workflows/Jules-Hotfix-Creator.yml
@@ -1,152 +1,0 @@
-name: Jules Hotfix-Creator (Worker)
-
-# Triggers when CI fails on protected branches (main/master)
-# Creates a hotfix branch, triggers Auto-Repair, and opens urgent PR
-
-on:
-  workflow_call:
-    inputs:
-      run_id:
-        required: true
-        type: string
-      failed_branch:
-        required: true
-        type: string
-
-permissions:
-  contents: write
-  pull-requests: write
-  actions: write
-
-jobs:
-  create-hotfix:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Validate failed branch
-        id: branch
-        run: |
-          FAILED_BRANCH='${{ inputs.failed_branch }}'
-          if [ -z "$FAILED_BRANCH" ]; then
-            echo "::error::Failed branch is empty."
-            exit 1
-          fi
-          if ! echo "$FAILED_BRANCH" | grep -Eq '^[A-Za-z0-9._/-]+$'; then
-            echo "::error::Failed branch contains invalid characters."
-            exit 1
-          fi
-          if [[ "$FAILED_BRANCH" == -* ]]; then
-            echo "::error::Failed branch must not start with '-'."
-            exit 1
-          fi
-          echo "failed_branch=$FAILED_BRANCH" >> "$GITHUB_OUTPUT"
-
-      - uses: actions/checkout@v4
-        with:
-          ref: ${{ steps.branch.outputs.failed_branch }}
-          fetch-depth: 0
-
-      - name: Validate Jules API Key
-        id: jules
-        env:
-          JULES_API_KEY: ${{ secrets.JULES_API_KEY }}
-        run: |
-          if [ -z "$JULES_API_KEY" ]; then
-            echo "::warning::JULES_API_KEY not configured. Skipping hotfix automation."
-            echo "enabled=false" >> $GITHUB_OUTPUT
-          else
-            echo "enabled=true" >> $GITHUB_OUTPUT
-          fi
-
-      - uses: actions/setup-node@v4
-        if: steps.jules.outputs.enabled == 'true'
-        with: { node-version: "20" }
-      - run: npm install -g @google/jules
-        if: steps.jules.outputs.enabled == 'true'
-      - run: jules auth --token ${{ secrets.JULES_API_KEY }}
-        if: steps.jules.outputs.enabled == 'true'
-
-      - name: Create Hotfix Branch
-        id: hotfix
-        if: steps.jules.outputs.enabled == 'true'
-        env:
-          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        run: |
-          TIMESTAMP=$(date +%Y%m%d-%H%M%S)
-          HOTFIX_BRANCH="hotfix/ci-fail-${TIMESTAMP}"
-
-          echo "Creating hotfix branch: $HOTFIX_BRANCH"
-          git checkout -b "$HOTFIX_BRANCH"
-          git push -u origin "$HOTFIX_BRANCH"
-
-          echo "branch=$HOTFIX_BRANCH" >> $GITHUB_OUTPUT
-
-      - name: Fetch CI Logs & Auto-Fix
-        if: steps.jules.outputs.enabled == 'true'
-        env:
-          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          TARGET_RUN_ID: ${{ inputs.run_id }}
-        run: |
-          HOTFIX_BRANCH="${{ steps.hotfix.outputs.branch }}"
-          FAILED_BRANCH="${{ steps.branch.outputs.failed_branch }}"
-          gh run view $TARGET_RUN_ID --log-failed > logs.txt
-
-          LOG_CONTENT=$(cat logs.txt)
-          PROMPT=$(printf '%s\n' \
-            "URGENT: CI failed on protected branch $FAILED_BRANCH." \
-            "Analyze the logs and create a minimal hotfix:" \
-            "1. **Formatting & Linting**: Run 'ruff check . --fix' and 'black .'" \
-            "2. **Import Sorting**: Run 'isort --profile black .'" \
-            "3. **Type Errors**: Add missing type hints or proper type ignores" \
-            "4. **Test Failures**: Fix failing tests (fix implementation, not tests)" \
-            "5. **Coverage**: Generate tests if coverage dropped below 80%" \
-            "6. **Syntax**: Fix any syntax errors or missing imports" \
-            "CRITICAL: This is a hotfix for $FAILED_BRANCH. Keep changes minimal and focused." \
-            "CI Failure Logs:" \
-            "$LOG_CONTENT")
-          jules task fix --branch "$HOTFIX_BRANCH" --prompt "$PROMPT"
-
-      - name: Create Urgent PR
-        if: steps.jules.outputs.enabled == 'true'
-        env:
-          GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-        run: |
-          HOTFIX_BRANCH="${{ steps.hotfix.outputs.branch }}"
-          FAILED_BRANCH="${{ steps.branch.outputs.failed_branch }}"
-
-          gh pr create \
-            --base "$FAILED_BRANCH" \
-            --head "$HOTFIX_BRANCH" \
-            --title "[Jules/Hotfix-Creator] 🚨 Urgent: Fix CI failure on $FAILED_BRANCH" \
-            --label "urgent,auto-fix,hotfix" \
-            --body "$(printf '%s\n' \
-            "## Automated Hotfix for Protected Branch" \
-            "" \
-            "**Failed Branch:** $FAILED_BRANCH" \
-            "**CI Run:** ${{ github.server_url }}/${{ github.repository }}/actions/runs/${{ inputs.run_id }}" \
-            "**Created By:** Jules Hotfix-Creator" \
-            "" \
-            "### Summary" \
-            "CI failed on the protected branch $FAILED_BRANCH. This automated hotfix attempts to resolve the issues." \
-            "" \
-            "### Changes" \
-            "Jules Auto-Repair has analyzed the CI failure logs and applied fixes for:" \
-            "- Code formatting and linting issues" \
-            "- Type checking errors" \
-            "- Test failures" \
-            "- Coverage gaps" \
-            "- Syntax errors" \
-            "" \
-            "### Review Checklist" \
-            "- [ ] Verify all CI checks pass" \
-            "- [ ] Review changes are minimal and focused" \
-            "- [ ] Confirm no unintended side effects" \
-            "- [ ] Merge immediately if changes look correct" \
-            "" \
-            "### Next Steps" \
-            "1. Review the changes in this PR" \
-            "2. If acceptable, merge immediately to restore ${{ inputs.failed_branch }}" \
-            "3. If issues remain, manually fix and push to this branch" \
-            "" \
-            "---" \
-            "This is an automated hotfix. Please review carefully before merging." \
-            )"
