--- .github/workflows/Jules-Sentinel.yml
@@ -1,167 +1,0 @@
-name: Jules Sentinel (Security Audit)
-
-on:
-  workflow_call:
-  workflow_dispatch:
-  # Schedule disabled - orchestrated by Jules-Control-Tower.yml
-  # schedule:
-  #   - cron: "0 0 * * 0,4"  # Weekly on Mondays at 1:30 AM UTC (overnight window)
-
-permissions:
-  contents: write
-  pull-requests: write
-  issues: write
-
-# HARDENING: Prevent concurrent runs
-concurrency:
-  group: jules-sentinel-${{ github.repository }}
-  cancel-in-progress: false
-
-jobs:
-  security-audit:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          fetch-depth: 0
-
-      - uses: actions/setup-python@v5
-        with:
-          python-version: "3.11"
-
-      - uses: actions/setup-node@v4
-        with:
-          node-version: "20"
-
-      - name: Install Dependencies
-        run: |
-          pip install safety bandit pip-audit semgrep || true
-          npm install -g @google/jules
-
-      - name: Run Security Scanners
-        run: |
-          mkdir -p reports
-
-          echo "Running pip-audit..."
-          pip-audit --format json -o reports/pip_audit.json 2>/dev/null || echo "[]" > reports/pip_audit.json
-
-          echo "Running bandit..."
-          bandit -r . -f json -o reports/bandit.json --exclude './.venv,./venv,./env,./.git' 2>/dev/null || echo "[]" > reports/bandit.json
-
-          echo "Running semgrep..."
-          semgrep --config=auto --json -o reports/semgrep.json . 2>/dev/null || echo "{}" > reports/semgrep.json
-
-      - name: Jules Sentinel Analysis
-        env:
-          JULES_API_KEY: ${{ secrets.JULES_API_KEY }}
-        run: |
-          set -e
-          DATE=$(date +%Y-%m-%d)
-          REPO="${{ github.repository }}"
-          API_BASE="https://jules.googleapis.com/v1alpha"
-
-          # Check if API key is configured
-          if [ -z "$JULES_API_KEY" ]; then
-            echo "::warning::JULES_API_KEY not configured. Skipping Jules task."
-            exit 0
-          fi
-
-          echo "Looking up repository source..."
-          SOURCES=$(curl -sf -H "X-Goog-Api-Key: $JULES_API_KEY" "$API_BASE/sources" || echo '{"sources":[]}')
-
-          OWNER=$(echo "$REPO" | cut -d'/' -f1)
-          REPO_NAME=$(echo "$REPO" | cut -d'/' -f2)
-
-          SOURCE_ID=$(echo "$SOURCES" | jq -r ".sources[] | select(.githubRepo.owner == \"$OWNER\" and .githubRepo.repo == \"$REPO_NAME\") | .name" | head -1)
-
-          if [ -z "$SOURCE_ID" ] || [ "$SOURCE_ID" = "null" ]; then
-            echo "::warning::Repository $REPO not found in Jules sources."
-            exit 0
-          fi
-
-          echo "Found source: $SOURCE_ID"
-
-          # Create session prompt
-          PROMPT=$(cat << 'PROMPT_EOF'
-          '
-          You are the SENTINEL security agent.
-
-          Analyze the security audit reports in ./reports/ and document findings.
-
-          ## DO NOT MODIFY CODE directly. Instead:
-
-          ### For HIGH severity: Create GitHub Issue with labels: security, high-priority, jules:sentinel
-          ### For MEDIUM severity: Add to .jules/sentinel.md and create Issue
-          ### For LOW severity: Document in .jules/sentinel.md only
-
-          Update .jules/sentinel.md with scan results summary.
-
-          CRITICAL: This is a READ-ONLY audit. Do not fix code - only document and create issues.
-          PROMPT_EOF
-          )
-
-          echo "Creating Jules session..."
-          # HARDENING: Do NOT use automationMode: AUTO_CREATE_PR
-          # Sentinel is read-only - it documents findings but doesn't create PRs autonomously
-          # This prevents cascade issues where Jules creates PRs that trigger more workflows
-          REQUEST_JSON=$(jq -n \
-            --arg prompt "$PROMPT" \
-            --arg source "$SOURCE_ID" \
-            '{
-              prompt: $prompt,
-              sourceContext: {
-                source: $source,
-                githubRepoContext: {
-                  startingBranch: "main"
-                }
-              }
-            }')
-
-          echo "Request: $REQUEST_JSON"
-
-          HTTP_CODE=$(curl -s -o /tmp/session_response.json -w "%{http_code}" -X POST \
-            -H "X-Goog-Api-Key: $JULES_API_KEY" \
-            -H "Content-Type: application/json" \
-            -d "$REQUEST_JSON" \
-            "$API_BASE/sessions")
-
-          SESSION_RESPONSE=$(cat /tmp/session_response.json)
-          echo "HTTP Status: $HTTP_CODE"
-
-          if [ "$HTTP_CODE" != "200" ]; then
-            echo "::error::Jules API returned HTTP $HTTP_CODE"
-            exit 1
-          fi
-
-          SESSION_ID=$(echo "$SESSION_RESPONSE" | jq -r '.name')
-          echo "Created session: $SESSION_ID"
-
-          # Poll for completion
-          MAX_WAIT=1800
-          POLL_INTERVAL=30
-          ELAPSED=0
-
-          while [ $ELAPSED -lt $MAX_WAIT ]; do
-            SESSION_STATUS=$(curl -sf -H "X-Goog-Api-Key: $JULES_API_KEY" "$API_BASE/$SESSION_ID")
-            STATE=$(echo "$SESSION_STATUS" | jq -r '.state')
-            echo "Session state: $STATE (elapsed: ${ELAPSED}s)"
-
-            case "$STATE" in
-              "COMPLETED"|"SUBMITTED")
-                echo "Session completed successfully!"
-                break
-                ;;
-              "FAILED"|"CANCELLED")
-                echo "::error::Session $STATE"
-                exit 1
-                ;;
-            esac
-            sleep $POLL_INTERVAL
-            ELAPSED=$((ELAPSED + POLL_INTERVAL))
-          done
-
-      - name: Commit Journal
-        uses: stefanzweifel/git-auto-commit-action@v5
-        with:
-          commit_message: "docs(sentinel): weekly security audit"
-          file_pattern: ".jules/sentinel.md"
