+++ .github/workflows/Manual-Run-All.yml
@@ -27,7 +27,7 @@ jobs:
   dispatch-workflows:
     runs-on: ubuntu-latest
     steps:
-      - name: Dispatch Assessment Generator
+      - name: Dispatch Comprehensive Assessment
         if: inputs.workflows == 'all' || inputs.workflows == 'assessments-only'
         uses: actions/github-script@v7
         with:
@@ -35,16 +35,16 @@ jobs:
             await github.rest.actions.createWorkflowDispatch({
               owner: context.repo.owner,
               repo: context.repo.repo,
-              workflow_id: 'Jules-Assessment-Generator.yml',
+              workflow_id: 'Jules-Comprehensive-Assessment.yml',
               ref: 'main'
             });
-            console.log('Dispatched Assessment Generator');
+            console.log('Dispatched Comprehensive Assessment');
 
       - name: Wait for Assessment
         if: inputs.workflows == 'all' || inputs.workflows == 'assessments-only'
         run: sleep 10
 
-      - name: Dispatch Code Quality Reviewer
+      - name: Dispatch Documentation Auditor
         if: inputs.workflows == 'all' || inputs.workflows == 'assessments-only'
         uses: actions/github-script@v7
         with:
@@ -52,64 +52,13 @@ jobs:
             await github.rest.actions.createWorkflowDispatch({
               owner: context.repo.owner,
               repo: context.repo.repo,
-              workflow_id: 'Jules-Code-Quality-Reviewer.yml',
+              workflow_id: 'Jules-Documentation-Auditor.yml',
               ref: 'main'
             });
-            console.log('Dispatched Code Quality Reviewer');
+            console.log('Dispatched Documentation Auditor');
 
-      - name: Wait for Quality Review
-        if: inputs.workflows == 'all' || inputs.workflows == 'assessments-only'
-        run: sleep 10
-
-      - name: Dispatch Completist
-        if: inputs.workflows == 'all' || inputs.workflows == 'assessments-only'
-        uses: actions/github-script@v7
-        with:
-          script: |
-            await github.rest.actions.createWorkflowDispatch({
-              owner: context.repo.owner,
-              repo: context.repo.repo,
-              workflow_id: 'Jules-Completist.yml',
-              ref: 'main'
-            });
-            console.log('Dispatched Completist');
-
-      - name: Wait for Completist
-        if: inputs.workflows == 'all' || inputs.workflows == 'assessments-only'
-        run: sleep 10
-
-      - name: Dispatch Sentinel
-        if: inputs.workflows == 'all' || inputs.workflows == 'assessments-only'
-        uses: actions/github-script@v7
-        with:
-          script: |
-            await github.rest.actions.createWorkflowDispatch({
-              owner: context.repo.owner,
-              repo: context.repo.repo,
-              workflow_id: 'Jules-Sentinel.yml',
-              ref: 'main'
-            });
-            console.log('Dispatched Sentinel');
-
-      - name: Wait for Sentinel
+      - name: Wait for Documentation Auditor
         if: inputs.workflows == 'all'
-        run: sleep 10
-
-      - name: Dispatch Issue Resolver
-        if: inputs.workflows == 'all' || inputs.workflows == 'fixes-only'
-        uses: actions/github-script@v7
-        with:
-          script: |
-            await github.rest.actions.createWorkflowDispatch({
-              owner: context.repo.owner,
-              repo: context.repo.repo,
-              workflow_id: 'Jules-Issue-Resolver.yml',
-              ref: 'main'
-            });
-            console.log('Dispatched Issue Resolver');
-
-      - name: Wait for Issue Resolver
-        if: inputs.workflows == 'all' || inputs.workflows == 'fixes-only'
         run: sleep 10
 
       - name: Dispatch PR Compiler
