+++ src/tools/model_generation/editor/__init__.py
@@ -7,14 +7,18 @@ This module provides editors for URDF manipulation:
 """
 
 from model_generation.editor.frankenstein_editor import (
-    ComponentReference,
-    ComponentType,
     FrankensteinEditor,
 )
+from model_generation.editor.frankenstein_types import (
+    ComponentReference,
+    ComponentType,
+)
 from model_generation.editor.text_editor import (
+    URDFTextEditor,
+)
+from model_generation.editor.text_editor_types import (
     DiffHunk,
     DiffResult,
-    URDFTextEditor,
     ValidationMessage,
     ValidationSeverity,
 )
