+++ src/tools/model_generation/editor/text_editor.py
@@ -6,98 +6,49 @@ Provides text-based editing of URDF files with:
 - Diff generation between versions
 - Undo/redo support
 - Real-time validation feedback
+
+Decomposed via SRP into:
+- text_editor_types.py: Shared data types
+- text_editor_validation.py: ValidationMixin (XML/URDF validation)
+- text_editor_diff.py: DiffMixin (diff computation)
 """
 
 from __future__ import annotations
 
-import difflib
 import hashlib
 import logging
 import re
 import xml.etree.ElementTree as ET
 from collections.abc import Callable
-from dataclasses import dataclass, field
 from datetime import datetime
-from enum import Enum
 from pathlib import Path
 from typing import Any
 
 import defusedxml.ElementTree as DefusedET
 
+from .text_editor_diff import DiffMixin
+from .text_editor_types import (
+    DiffHunk,
+    DiffResult,
+    EditorVersion,
+    ValidationMessage,
+    ValidationSeverity,
+)
+from .text_editor_validation import ValidationMixin
+
 logger = logging.getLogger(__name__)
 
-
-class ValidationSeverity(Enum):
-    """Severity levels for validation messages."""
-
-    ERROR = "error"
-    WARNING = "warning"
-    INFO = "info"
-
-
-@dataclass
-class ValidationMessage:
-    """A validation message for URDF content."""
-
-    severity: ValidationSeverity
-    line: int
-    column: int
-    message: str
-    element: str | None = None
-
-    def __str__(self) -> str:
-        prefix = self.severity.value.upper()
-        loc = f"Line {self.line}"
-        if self.column > 0:
-            loc += f", Col {self.column}"
-        if self.element:
-            return f"[{prefix}] {loc} ({self.element}): {self.message}"
-        return f"[{prefix}] {loc}: {self.message}"
+# Re-export types for backwards compatibility
+__all__ = [
+    "DiffHunk",
+    "DiffResult",
+    "URDFTextEditor",
+    "ValidationMessage",
+    "ValidationSeverity",
+]
 
 
-@dataclass
-class DiffHunk:
-    """A single hunk in a diff."""
-
-    old_start: int
-    old_count: int
-    new_start: int
-    new_count: int
-    lines: list[str]
-    context_before: list[str] = field(default_factory=list)
-    context_after: list[str] = field(default_factory=list)
-
-
-@dataclass
-class DiffResult:
-    """Result of comparing two URDF versions."""
-
-    original_content: str
-    modified_content: str
-    hunks: list[DiffHunk]
-    unified_diff: str
-    additions: int
-    deletions: int
-    has_changes: bool
-
-    def get_summary(self) -> str:
-        """Get a summary of changes."""
-        if not self.has_changes:
-            return "No changes"
-        return f"{self.additions} additions, {self.deletions} deletions in {len(self.hunks)} hunks"
-
-
-@dataclass
-class EditorVersion:
-    """A version of the document."""
-
-    content: str
-    timestamp: datetime
-    description: str
-    checksum: str
-
-
-class URDFTextEditor:
+class URDFTextEditor(ValidationMixin, DiffMixin):
     """
     Text editor for URDF files with validation and diff support.
 
@@ -345,566 +296,6 @@ class URDFTextEditor:
 
         content = self._content.replace(old_content, new_content, 1)
         return self.set_content(content, f"Replace {element_name}")
-
-    # ============================================================
-    # Validation
-    # ============================================================
-
-    def validate(self) -> list[ValidationMessage]:
-        """
-        Validate current URDF content.
-
-        Returns:
-            List of validation messages
-        """
-        messages = []
-
-        # XML validation
-        messages.extend(self._validate_xml())
-
-        if not any(m.severity == ValidationSeverity.ERROR for m in messages):
-            # URDF-specific validation
-            messages.extend(self._validate_urdf())
-
-        return messages
-
-    def _validate_xml(self) -> list[ValidationMessage]:
-        """Validate XML syntax."""
-        messages = []
-
-        try:
-            DefusedET.fromstring(self._content)
-        except ET.ParseError as e:
-            # Parse error message for line/column
-            error_str = str(e)
-            line, col = 1, 0
-
-            # Try to extract line number
-            match = re.search(r"line (\d+)", error_str)
-            if match:
-                line = int(match.group(1))
-
-            match = re.search(r"column (\d+)", error_str)
-            if match:
-                col = int(match.group(1))
-
-            messages.append(
-                ValidationMessage(
-                    severity=ValidationSeverity.ERROR,
-                    line=line,
-                    column=col,
-                    message=f"XML syntax error: {error_str}",
-                )
-            )
-
-        return messages
-
-    def _validate_urdf(self) -> list[ValidationMessage]:
-        """Validate URDF-specific rules."""
-        messages: list[ValidationMessage] = []
-
-        try:
-            root = DefusedET.fromstring(self._content)
-        except ET.ParseError:
-            return messages  # Already reported in XML validation
-
-        # Check root element
-        if root.tag != "robot":
-            messages.append(
-                ValidationMessage(
-                    severity=ValidationSeverity.ERROR,
-                    line=1,
-                    column=0,
-                    message=f"Root element should be 'robot', got '{root.tag}'",
-                )
-            )
-            return messages
-
-        # Check robot name
-        if not root.get("name"):
-            messages.append(
-                ValidationMessage(
-                    severity=ValidationSeverity.WARNING,
-                    line=1,
-                    column=0,
-                    message="Robot element missing 'name' attribute",
-                    element="robot",
-                )
-            )
-
-        # Collect and validate links and joints
-        links = self._validate_urdf_links(root, messages)
-        joints = self._validate_urdf_joints(root, links, messages)
-
-        # Check for orphan links (no joint connection)
-        self._validate_urdf_orphan_links(root, links, joints, messages)
-
-        return messages
-
-    def _validate_urdf_links(
-        self,
-        root: ET.Element,
-        messages: list[ValidationMessage],
-    ) -> dict[str, ET.Element]:
-        """Validate all link elements and return a map of name to element."""
-        links: dict[str, ET.Element] = {}
-
-        for _idx, link_elem in enumerate(root.findall("link")):
-            name = link_elem.get("name")
-            if not name:
-                line = self._find_element_line(link_elem)
-                messages.append(
-                    ValidationMessage(
-                        severity=ValidationSeverity.ERROR,
-                        line=line,
-                        column=0,
-                        message="Link element missing 'name' attribute",
-                        element="link",
-                    )
-                )
-            elif name in links:
-                line = self._find_element_line(link_elem)
-                messages.append(
-                    ValidationMessage(
-                        severity=ValidationSeverity.ERROR,
-                        line=line,
-                        column=0,
-                        message=f"Duplicate link name: '{name}'",
-                        element=name,
-                    )
-                )
-            else:
-                links[name] = link_elem
-
-            # Check inertial mass
-            self._validate_link_inertial(link_elem, name, messages)
-
-        return links
-
-    def _validate_link_inertial(
-        self,
-        link_elem: ET.Element,
-        link_name: str | None,
-        messages: list[ValidationMessage],
-    ) -> None:
-        """Validate the inertial/mass properties of a single link element."""
-        inertial = link_elem.find("inertial")
-        if inertial is None:
-            return
-
-        mass_elem = inertial.find("mass")
-        if mass_elem is None:
-            return
-
-        mass = mass_elem.get("value")
-        if mass is None:
-            return
-
-        try:
-            mass_val = float(mass)
-            if mass_val < 0:
-                line = self._find_element_line(mass_elem)
-                messages.append(
-                    ValidationMessage(
-                        severity=ValidationSeverity.ERROR,
-                        line=line,
-                        column=0,
-                        message=f"Negative mass value: {mass_val}",
-                        element=link_name,
-                    )
-                )
-            elif mass_val == 0:
-                line = self._find_element_line(mass_elem)
-                messages.append(
-                    ValidationMessage(
-                        severity=ValidationSeverity.WARNING,
-                        line=line,
-                        column=0,
-                        message="Zero mass value",
-                        element=link_name,
-                    )
-                )
-        except ValueError:
-            line = self._find_element_line(mass_elem)
-            messages.append(
-                ValidationMessage(
-                    severity=ValidationSeverity.ERROR,
-                    line=line,
-                    column=0,
-                    message=f"Invalid mass value: '{mass}'",
-                    element=link_name,
-                )
-            )
-
-    def _validate_urdf_joints(
-        self,
-        root: ET.Element,
-        links: dict[str, ET.Element],
-        messages: list[ValidationMessage],
-    ) -> dict[str, ET.Element]:
-        """Validate all joint elements and return a map of name to element."""
-        joints: dict[str, ET.Element] = {}
-
-        for _idx, joint_elem in enumerate(root.findall("joint")):
-            name = joint_elem.get("name")
-            if not name:
-                line = self._find_element_line(joint_elem)
-                messages.append(
-                    ValidationMessage(
-                        severity=ValidationSeverity.ERROR,
-                        line=line,
-                        column=0,
-                        message="Joint element missing 'name' attribute",
-                        element="joint",
-                    )
-                )
-            elif name in joints:
-                line = self._find_element_line(joint_elem)
-                messages.append(
-                    ValidationMessage(
-                        severity=ValidationSeverity.ERROR,
-                        line=line,
-                        column=0,
-                        message=f"Duplicate joint name: '{name}'",
-                        element=name,
-                    )
-                )
-            else:
-                joints[name] = joint_elem
-
-            # Check joint type
-            joint_type = joint_elem.get("type")
-            valid_types = {
-                "revolute",
-                "continuous",
-                "prismatic",
-                "fixed",
-                "floating",
-                "planar",
-            }
-            if joint_type not in valid_types:
-                line = self._find_element_line(joint_elem)
-                messages.append(
-                    ValidationMessage(
-                        severity=ValidationSeverity.ERROR,
-                        line=line,
-                        column=0,
-                        message=f"Invalid joint type: '{joint_type}'",
-                        element=name,
-                    )
-                )
-
-            # Check parent/child references and limits
-            self._validate_joint_references(
-                joint_elem, name, joint_type, links, messages
-            )
-
-        return joints
-
-    def _validate_joint_references(
-        self,
-        joint_elem: ET.Element,
-        name: str | None,
-        joint_type: str | None,
-        links: dict[str, ET.Element],
-        messages: list[ValidationMessage],
-    ) -> None:
-        """Validate parent/child references and limits for a single joint."""
-        parent_elem = joint_elem.find("parent")
-        child_elem = joint_elem.find("child")
-
-        if parent_elem is None:
-            line = self._find_element_line(joint_elem)
-            messages.append(
-                ValidationMessage(
-                    severity=ValidationSeverity.ERROR,
-                    line=line,
-                    column=0,
-                    message="Joint missing parent element",
-                    element=name,
-                )
-            )
-        else:
-            parent_link = parent_elem.get("link")
-            if parent_link and parent_link not in links:
-                line = self._find_element_line(parent_elem)
-                messages.append(
-                    ValidationMessage(
-                        severity=ValidationSeverity.ERROR,
-                        line=line,
-                        column=0,
-                        message=f"Parent link not found: '{parent_link}'",
-                        element=name,
-                    )
-                )
-
-        if child_elem is None:
-            line = self._find_element_line(joint_elem)
-            messages.append(
-                ValidationMessage(
-                    severity=ValidationSeverity.ERROR,
-                    line=line,
-                    column=0,
-                    message="Joint missing child element",
-                    element=name,
-                )
-            )
-        else:
-            child_link = child_elem.get("link")
-            if child_link and child_link not in links:
-                line = self._find_element_line(child_elem)
-                messages.append(
-                    ValidationMessage(
-                        severity=ValidationSeverity.ERROR,
-                        line=line,
-                        column=0,
-                        message=f"Child link not found: '{child_link}'",
-                        element=name,
-                    )
-                )
-
-        # Check limits for revolute/prismatic
-        if joint_type in {"revolute", "prismatic"}:
-            limit_elem = joint_elem.find("limit")
-            if limit_elem is None:
-                line = self._find_element_line(joint_elem)
-                messages.append(
-                    ValidationMessage(
-                        severity=ValidationSeverity.WARNING,
-                        line=line,
-                        column=0,
-                        message=f"{joint_type} joint missing limit element",
-                        element=name,
-                    )
-                )
-
-    def _validate_urdf_orphan_links(
-        self,
-        root: ET.Element,
-        links: dict[str, ET.Element],
-        joints: dict[str, ET.Element],
-        messages: list[ValidationMessage],
-    ) -> None:
-        """Check for orphan links that have no joint connection."""
-        child_links: set[str | None] = set()
-        for joint_elem in root.findall("joint"):
-            child_elem = joint_elem.find("child")
-            if child_elem is not None:
-                child_links.add(child_elem.get("link"))
-
-        for link_name in links:
-            if link_name not in child_links:
-                # This might be the root link
-                is_parent = any(
-                    (parent := j.find("parent")) is not None
-                    and parent.get("link") == link_name
-                    for j in root.findall("joint")
-                )
-                if not is_parent and len(links) > 1:
-                    messages.append(
-                        ValidationMessage(
-                            severity=ValidationSeverity.WARNING,
-                            line=1,
-                            column=0,
-                            message=f"Link '{link_name}' is not connected to any joint",
-                            element=link_name,
-                        )
-                    )
-
-    def _find_element_line(self, elem: ET.Element) -> int:
-        """Find the line number of an element (approximate)."""
-        # This is a simple heuristic - search for element in content
-        ET.tostring(elem, encoding="unicode")
-        tag_start = f"<{elem.tag}"
-
-        # Find in content
-        lines = self._content.split("\n")
-        for idx, line in enumerate(lines, 1):
-            if tag_start in line:
-                # Check if attributes match
-                name = elem.get("name")
-                if name is None or f'name="{name}"' in line or f"name='{name}'" in line:
-                    return idx
-
-        return 1
-
-    # ============================================================
-    # Diff Operations
-    # ============================================================
-
-    def get_diff_from_original(self) -> DiffResult:
-        """
-        Get diff between current content and original.
-
-        Returns:
-            DiffResult with changes
-        """
-        return self._compute_diff(self._original_content, self._content)
-
-    def get_diff_between_versions(
-        self,
-        version_a: int,
-        version_b: int,
-    ) -> DiffResult:
-        """
-        Get diff between two versions in history.
-
-        Args:
-            version_a: First version index
-            version_b: Second version index
-
-        Returns:
-            DiffResult with changes
-        """
-        if version_a < 0 or version_a >= len(self._history):
-            raise IndexError(f"Invalid version index: {version_a}")
-        if version_b < 0 or version_b >= len(self._history):
-            raise IndexError(f"Invalid version index: {version_b}")
-
-        content_a = self._history[version_a].content
-        content_b = self._history[version_b].content
-
-        return self._compute_diff(content_a, content_b)
-
-    def get_diff_with_string(self, other_content: str) -> DiffResult:
-        """
-        Get diff between current content and provided string.
-
-        Args:
-            other_content: Content to compare with
-
-        Returns:
-            DiffResult with changes
-        """
-        return self._compute_diff(self._content, other_content)
-
-    def _compute_diff(self, original: str, modified: str) -> DiffResult:
-        """Compute diff between two strings."""
-        original_lines = original.splitlines(keepends=True)
-        modified_lines = modified.splitlines(keepends=True)
-
-        # Generate unified diff
-        diff_lines = list(
-            difflib.unified_diff(
-                original_lines,
-                modified_lines,
-                fromfile="original",
-                tofile="modified",
-                lineterm="",
-            )
-        )
-        unified_diff = "".join(diff_lines)
-
-        # Parse hunks
-        hunks: list[DiffHunk] = []
-        current_hunk_lines: list[str] = []
-        old_start, old_count, new_start, new_count = 0, 0, 0, 0
-        additions = 0
-        deletions = 0
-
-        for line in diff_lines:
-            if line.startswith("@@"):
-                # Save previous hunk
-                if current_hunk_lines:
-                    hunks.append(
-                        DiffHunk(
-                            old_start=old_start,
-                            old_count=old_count,
-                            new_start=new_start,
-                            new_count=new_count,
-                            lines=current_hunk_lines,
-                        )
-                    )
-                    current_hunk_lines = []
-
-                # Parse hunk header
-                match = re.match(r"@@ -(\d+)(?:,(\d+))? \+(\d+)(?:,(\d+))? @@", line)
-                if match:
-                    old_start = int(match.group(1))
-                    old_count = int(match.group(2) or 1)
-                    new_start = int(match.group(3))
-                    new_count = int(match.group(4) or 1)
-            elif line.startswith(("---", "+++")):
-                continue
-            elif line.startswith("+"):
-                current_hunk_lines.append(line)
-                additions += 1
-            elif line.startswith("-"):
-                current_hunk_lines.append(line)
-                deletions += 1
-            elif line.startswith(" ") or line == "\n":
-                current_hunk_lines.append(line)
-
-        # Save last hunk
-        if current_hunk_lines:
-            hunks.append(
-                DiffHunk(
-                    old_start=old_start,
-                    old_count=old_count,
-                    new_start=new_start,
-                    new_count=new_count,
-                    lines=current_hunk_lines,
-                )
-            )
-
-        return DiffResult(
-            original_content=original,
-            modified_content=modified,
-            hunks=hunks,
-            unified_diff=unified_diff,
-            additions=additions,
-            deletions=deletions,
-            has_changes=original != modified,
-        )
-
-    def get_side_by_side_diff(
-        self,
-        original: str | None = None,
-        modified: str | None = None,
-        context_lines: int = 3,
-    ) -> list[tuple[str | None, str | None, str]]:
-        """
-        Get side-by-side diff representation.
-
-        Args:
-            original: Original content (default: original file content)
-            modified: Modified content (default: current content)
-            context_lines: Number of context lines
-
-        Returns:
-            List of (left_line, right_line, change_type) tuples.
-            change_type is one of: 'equal', 'insert', 'delete', 'replace'
-        """
-        if original is None:
-            original = self._original_content
-        if modified is None:
-            modified = self._content
-
-        original_lines = original.splitlines()
-        modified_lines = modified.splitlines()
-
-        differ = difflib.SequenceMatcher(None, original_lines, modified_lines)
-
-        result: list[tuple[str | None, str | None, str]] = []
-        for opcode, i1, i2, j1, j2 in differ.get_opcodes():
-            if opcode == "equal":
-                for i, j in zip(range(i1, i2), range(j1, j2), strict=False):
-                    result.append((original_lines[i], modified_lines[j], "equal"))
-            elif opcode == "insert":
-                for j in range(j1, j2):
-                    result.append((None, modified_lines[j], "insert"))
-            elif opcode == "delete":
-                for i in range(i1, i2):
-                    result.append((original_lines[i], None, "delete"))
-            elif opcode == "replace":
-                # Match up lines as best we can
-                max_len = max(i2 - i1, j2 - j1)
-                for k in range(max_len):
-                    left = original_lines[i1 + k] if i1 + k < i2 else None
-                    right = modified_lines[j1 + k] if j1 + k < j2 else None
-                    result.append((left, right, "replace"))
-
-        return result
 
     # ============================================================
     # History / Undo / Redo
